requirements.txt
Flask==2.3.3
flask-smorest==0.42.1
python-dotenv==1.0.0
openai==0.27.8
langchain==0.0.284
langgraph==0.0.55
sympy==1.12
graphviz==0.20.1
marshmallow==3.20.1
pytest==7.4.2
reportlab==4.0.4


requests~=2.32.5
typing_extensions~=4.15.0